Extract replace model specific elasticity laws with unified implementation
