+++ src/fenics_constitutive/models/mises_plasticity_isotropic_hardening.py
@@ -8,6 +8,8 @@ from fenics_constitutive import (
     strain_from_grad_u,
 )
 
+from .elasticity_laws import FullConstraintLaw
+
 
 class VonMises3D(IncrSmallStrainModel):
     r"""
@@ -43,12 +45,10 @@ class VonMises3D(IncrSmallStrainModel):
                 [0, 0, 0, 0, 0, 0],
             ]
         )
-        self.I2 = np.zeros(
-            self.stress_strain_dim, dtype=np.float64
-        )  # Identity of rank 2 tensor
-        self.I2[0] = 1.0
-        self.I2[1] = 1.0
-        self.I2[2] = 1.0
+
+        law = FullConstraintLaw()
+        self.I2 = law.get_I2(self.stress_strain_dim)
+
         self.I4 = np.eye(
             self.stress_strain_dim, dtype=np.float64
         )  # Identity of rank 4 tensor
